Split ParseExtensions into multiple files
