+++ Funcky/Extensions/ParseExtensions/ParseExtensions.Numbers.cs
@@ -6,10 +6,6 @@ namespace Funcky.Extensions
     public static partial class ParseExtensions
     {
         [Pure]
-        public static Option<bool> ParseBooleanOrNone(this string candidate)
-            => FailToOption<bool>.FromTryPattern(bool.TryParse, candidate);
-
-        [Pure]
         public static Option<byte> ParseByteOrNone(this string candidate)
             => FailToOption<byte>.FromTryPattern(byte.TryParse, candidate);
 
@@ -64,31 +60,5 @@ namespace Funcky.Extensions
         [Pure]
         public static Option<decimal> ParseDecimalOrNone(this string candidate, NumberStyles styles, IFormatProvider provider)
             => FailToOption<decimal>.FromTryPattern(decimal.TryParse, candidate, styles, provider);
-
-        [Pure]
-        public static Option<DateTime> ParseDateTimeOrNone(this string candidate)
-            => FailToOption<DateTime>.FromTryPattern(DateTime.TryParse, candidate);
-
-        [Pure]
-        public static Option<DateTime> ParseDateTimeOrNone(this string candidate, IFormatProvider provider, DateTimeStyles styles)
-            => FailToOption<DateTime>.FromTryPattern(DateTime.TryParse, candidate, provider, styles);
-
-        [Pure]
-        public static Option<TimeSpan> ParseTimeSpanOrNone(this string candidate)
-            => FailToOption<TimeSpan>.FromTryPattern(TimeSpan.TryParse, candidate);
-
-        [Pure]
-        public static Option<TimeSpan> ParseTimeSpanOrNone(this string candidate, IFormatProvider provider)
-            => FailToOption<TimeSpan>.FromTryPattern(TimeSpan.TryParse, candidate, provider);
-
-        [Pure]
-        public static Option<TEnum> ParseEnumOrNone<TEnum>(this string candidate)
-            where TEnum : struct
-            => FailToOption<TEnum>.FromTryPattern(Enum.TryParse, candidate);
-
-        [Pure]
-        public static Option<TEnum> ParseEnumOrNone<TEnum>(this string candidate, bool ignoreCase)
-            where TEnum : struct
-            => FailToOption<TEnum>.FromTryPattern(Enum.TryParse, candidate, ignoreCase);
     }
 }
